Can't fix the mac rcmd CHECK in GitHub actions, see https://github.com/lcolladotor/biocthis/issues/27
so changed to `error=TRUE` in RMarkdown doc for `ceef.Rmd` to see if this will just get it to pass.
This seems to be something to do with fonts or trying to plot outside the plotting area. The log says its using UTF-8 so perhaps need to explicityl say this beforehand? The vignettes all build fine inside RStudio including using `devtools::check_mac_version()`